+++ cmd/orgs.go
@@ -5,9 +5,7 @@ import (
 	"fmt"
 	"os"
 	"sort"
-	"strconv"
 	"sync"
-	//"text/tabwriter"
 
 	"github.com/urfave/cli"
 	"github.com/juju/ansiterm"
@@ -137,7 +135,7 @@ func orgsListCmd(ctx *cli.Context) error {
 	if session.Type() == apitypes.UserSession {
 		for i, o := range orgs {
 			if o.Body.Name == session.Username() {
-				fmt.Printf("%s %s\n", o.Body.Name, "(" + ui.Color(ansiterm.DarkGray, "personal") + ")")
+				fmt.Printf("%s %s\n", o.Body.Name, "(" + ui.Faint("personal") + ")")
 				withoutPersonal = append(orgs[:i], orgs[i+1:]...)
 			}
 		}
@@ -393,7 +391,7 @@ func orgsMembersListCmd(ctx *cli.Context) error {
 	for _, user := range users {
 		me := ""
 		if session.Username() == user.Body.Username {
-			me = ui.Color(ansiterm.DarkGray, "*")
+			me = ui.Faint("*")
 		}
 
 		// Sort teams by precedence
@@ -416,17 +414,8 @@ func orgsMembersListCmd(ctx *cli.Context) error {
 
 	w.Flush()
 
-	count := strconv.Itoa(len(userIDs))
-	var countStr string
-	if len(userIDs) == 1 {
-		countStr = "org " + org.Body.Name + " has (" + count + ") member."
-	} else {
-		countStr = "org " + org.Body.Name + " has (" + count + ") members."
-	}
+	fmt.Printf("\nOrg %s has (%d) member%s\n", org.Body.Name, len(userIDs), plural(len(userIDs)))
 
-	fmt.Println("")
-	fmt.Println(countStr)
-	fmt.Println("")
 	return nil
 }
 
@@ -437,6 +426,45 @@ func getOrg(ctx context.Context, client *api.Client, name string) (*envelope.Org
 	}
 	if org == nil {
 		return nil, errs.NewExitError("Org not found.")
+	}
+
+	return org, nil
+}
+
+// This functions is intended to be used when a command takes an optional org flag.
+// In the situation where no org is provided in the flags, getOrgWithPrompt prompts the user
+// to select from a list of exisitng orgs, using SelectExistingOrgPrompt.
+// In the situation where an org is provided in the flags, the function returns the associated
+// org.Envelope structure.
+func getOrgWithPrompt(client *api.Client, c context.Context, orgName string) (*envelope.Org, error) {
+
+	var org *envelope.Org
+
+	if orgName == "" {
+		// Retrieve list of available orgs
+		orgs, err := client.Orgs.List(c)
+		if err != nil {
+			return nil, err
+		}
+
+		// Prompt user to select from list of existing orgs
+		idx, _, err := SelectExistingOrgPrompt(orgs)
+		if err != nil {
+			return nil, err
+		}
+
+		org = &orgs[idx]
+
+	} else {
+		// If org flag was used, identify the org supplied.
+		var err error
+		org, err = client.Orgs.GetByName(c, orgName)
+		if err != nil {
+			return nil, err
+		}
+		if org == nil {
+			return nil, errs.NewExitError("org " + orgName + " not found.")
+		}
 	}
 
 	return org, nil