+++ ui/color.go
@@ -6,14 +6,6 @@ import (
 	"github.com/juju/ansiterm"
 )
 
-var (
-  ServiceColor      = ansiterm.Magenta
-  EnvironmentColor  = ansiterm.BrightRed
-  CredPathColor     = ansiterm.DarkGray
-)
-
-var CredStyle       = ansiterm.Bold
-
 func Bold(s string) string {
   ctx := ansiterm.Context {
     Styles: []ansiterm.Style{ansiterm.Bold},
@@ -22,28 +14,16 @@ func Bold(s string) string {
   return createStyledString(ctx, s)
 }
 
+func Faint(s string) string {
+  return Color(ansiterm.DarkGray, s)
+}
+
 func Color(c ansiterm.Color, s string) string {
   ctx := ansiterm.Context {
     Foreground: c,
   }
 
   return createStyledString(ctx, s)
-}
-
-func Environment(s string) string {
-  return Color(EnvironmentColor, s)
-}
-
-func Service(s string) string {
-  return Color(ServiceColor, s)
-}
-
-func Cred(s string) string {
-  return Bold(s)
-}
-
-func CredPath(s string) string {
-  return Color(CredPathColor, s)
 }
 
 func BoldColor(c ansiterm.Color, s string) string {